Refactor out types.ts
